chore(engine): Remove workflow admission lanes
The max_scan_task_parallelism YAML setting stays in the config surface for compatibility, but workflow preparation no longer reads it. Workflows now dispatch every task directly through the runner without maintaining admission-lane token buckets.
